bug fix in line

+++ src/shapes/circle.rs
@@ -16,10 +16,10 @@ impl Circle {
 
     fn get_circle_rect_area(&self, canvas: &mut impl Canvas) -> (usize, usize, usize, usize) {
         let (center_x, center_y, radius) = (self.cx as f64, self.cy as f64, self.radius as f64);
-        let x1 = canvas.clamp_col(center_x - radius) as usize;
-        let x2 = canvas.clamp_col(center_x + radius) as usize;
-        let y1 = canvas.clamp_row(center_y - radius) as usize;
-        let y2 = canvas.clamp_row(center_y + radius) as usize;
+        let x1 = canvas.clamp_col((center_x - radius) as i64) as usize;
+        let x2 = canvas.clamp_col((center_x + radius) as i64) as usize;
+        let y1 = canvas.clamp_row((center_y - radius) as i64) as usize;
+        let y2 = canvas.clamp_row((center_y + radius) as i64) as usize;
 
         (x1, x2, y1, y2)
     }
